Completed 32 bit adder
This adder was modified & tested, most values tested are aligned with the theoretical values. However, the testbench requires verification using a calculator, and only displays the random values chosen and the resulting sum. In addition, for some reason the value for Cin seems to always generate zero, even though it was set to choose a random binary value.